samclient prelim
