+++ src/main/scala/org/broadinstitute/dsde/workbench/leonardo/auth/SamClient.scala
@@ -1,0 +1,5 @@
+package org.broadinstitute.dsde.workbench.leonardo.auth
+
+abstract class SamClient {
+
+}
